Rename to mjs
Was having problems with the js extension. nyc was not working with `type: module` in package.json, so I had to remove it and change to using .mjs. I think it's better because people importing the library in their projects would end up having to changing their package.json and that may not be a possibility for them.