Replace Mocha with Jest
I was having too many problems with c8 for test coverage, and the documentation is not that helpful. It is not as popular as nyc, so it's very difficult to find blogs and tutorials explaining how to use it, especially with TypeScript.

That's why I am migrating to Jest, which has broader adoption and plenty of people talking about it. Porting to Jest was ridiculously easy and, since describe() and it() work out of the box, I didn't even have to change the Mocha code! Well, I ended up replacing it() with test() because it's nicer to read, but it was not necessary.