Fix timeout during release on slow machine
The timeout of 10s was not enough to run the integration tests on a slower machine. Here I increase the timeout to 60s and improve the code to display a message whenever a timeout occurs (it took me a while to understand that the problem was the timeout and the message will help if it happens again).